포트분리 + 폴더구조수정 + 결제가능
