+++ page_user/views/home/menulist.ejs
@@ -53,18 +53,19 @@
         area_nm: areaName,
       },
       success: function (data) {
-        console.log(data, data.length);
+        // console.log(data);
+        var result = JSON.parse(data);
+        // console.log(result);
 
         // API에서 휴게소 메뉴 못받아오면 에러처리
-        if (data.length == 0) {
+        if (result.length == 0) {
           alert('휴게소에 음식이 없습니다.');
           window.location.href = '/home';
         } else {
-          resultMenuList(data);
+          resultMenuList(result);
         }
       }
     });
-
 
     //API 가져와서 카드에 메뉴넣는 함수
     function resultMenuList(data) {
@@ -72,12 +73,12 @@
       $.each(data, function (key, value) {
         html += "<div class='panel panel-default'>";
         html += "<div class='panel-body'>";
-        html += "<div style='padding-left:20px;'> <h4>" + value.foodNm;
-        if (value.bestfoodyn == 'Y') {
+        html += "<div style='padding-left:20px;'> <h4>" + value.food_nm;
+        if (value.best_yn == 'Y') {
           html += "<i class='fas fa-star' style='color:#FDB612'></i>";
         }
         html += "</h4>";
-        html += "<p class='card-text' id=cost> " + value.foodCost + "원</p></div><br>";
+        html += "<p class='card-text' id=cost> " + value.food_price + "원</p></div><br>";
         html += "<div class='container'>";
         html += "<div class='row'>";
         html += `<div class="col-xs-6">
